% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{midi_close_all}
\alias{midi_close_all}
\title{Close all open MIDI watchers}
\usage{
midi_close_all()
}
\value{
Used for side-effect
}
\description{
Close all open MIDI watchers
}
\examples{
if (interactive()){
  midi_close_all()
}
}
