% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_midi.R
\name{midi_list}
\alias{midi_list}
\title{List MIDI interfaces}
\usage{
midi_list()
}
\value{
A JSON array of interface names
}
\description{
This function will list all available MIDI interfaces.
}
\examples{
if (interactive()){
  midi_list()
}
}
