% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm_install.R
\name{noon_npm_install}
\alias{noon_npm_install}
\title{Install noon npm dependencies}
\usage{
noon_npm_install(package = "noon", force = FALSE)
}
\arguments{
\item{package}{package name}

\item{force}{Prompt the user to accept install?}
}
\value{
Used for side effect
}
\description{
Install noon npm dependencies
}
\examples{
if (interactive()){
  noon_npm_install()
}
}
